Fixed issue with Travis-CI build timeout
xtext-idea-gradle-plugin DownloadIdea task processes
for extended period of time without standard output logging. As per
Travis-CI build guidelines, if a build exceeds 10 minutes without
log output, it will timeout and subsequently fail. A logging service
was added as an extension to the Gradle task to ensure repeated logging.

Caching of the ideaSDK directory was also added for subsequent Travis-CI
builds.