Added domain-specific query 'setter' behaviour
Within formatted queries made to the SWI-Prolog engine,
the syntax PREV_TERM <- NEW_TERM can be used to allow for
setting of term data stored in the Prolog knowledge base.

While this is an incubating feature (term definitions are
not robust and could potentially cause format issues), this
represents a major milestone in terms of added functionality
on top of JPL bringing Mochalog closer to approximating
object-oriented concepts in a Prolog context.