Added support for custom Jitpack builds
As Mochalog is not yet hosted in Maven Central, adding
Mochalog as a dependency is currently achieved through Jitpack
(https://jitpack.io/). Jitpack builds of Mochalog currently fail
due to SWI-Prolog not being present on the build environment,
hence custom Jitpack builds have been configured to skip tests
(which rely on SWI-Prolog functionality).

When Maven Central releases are facilitated, Jitpack will no
longer be needed. For the meantime, this is a robust solution
to allow software dependent on Mochalog to conveniently
fetch releases.